refactor: rename PrismaResaleRepository to PrismaResalesRepository and update import
